recommender/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.3.2
pymongo==4.8.2
python-dotenv==1.0.0
